requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
tensorflow==2.18.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
locust==2.31.8
loguru==0.7.2
httpx==0.27.2
pytest==8.3.3
pytest-cov==5.0.0
